+++ src/zigbee2mqtt.html
@@ -1,65 +1,3 @@
-<script type="text/javascript">
-    function refresh() {
-        RED.nodes.eachNode(function (node) {
-
-        });
-    }
-
-    RED.sidebar.addTab({
-        id: "zigbee2mqtt",
-        label: "Zigbee 2 MQTT",
-        name: "zigbee2mqtt",
-        content: "<div class=\"nr-db-sb\">TODO: Add overview of all devices</div>",
-        closeable: true,
-        pinned: true,
-        iconClass: "fa fa-cloud",
-        disableOnEdit: true,
-        onchange: function () { refresh(); }
-    });
-
-    if (RED.bavaria === undefined) {
-        RED.bavaria = {
-            validators: {
-                range: function (min, max) {
-                    return function (v) {
-                        return v <= max && v >= min;
-                    }
-                }
-            },
-            ui: {
-                addNumInput: function (element) {
-                    $(element).typedInput({
-                        type: "num",
-                        types: ["num"],
-                    });
-                },
-                addNumInputs: function (elements) {
-                    elements.forEach(RED.bavaria.ui.addNumInput);
-                },
-                addNumRangeInput: function (element, min, max) {
-                    RED.bavaria.ui.addNumInput(element);
-                    var validator = RED.bavaria.validators.range(min, max);
-
-                    $(element).change(function () {
-                        var valid = validator($(this).val());
-                        var targetElement = $(this).parent().find('.red-ui-typedInput-container');
-                        if (!valid) {
-                            targetElement.addClass('input-error');
-                        } else {
-                            targetElement.removeClass('input-error');
-                        }
-                    });
-                },
-                addNumRangeInputs: function (elements, min, max) {
-                    elements.forEach(element => {
-                        RED.bavaria.ui.addNumRangeInput(element, min, max);
-                    });
-                }
-            }
-        }
-    }
-</script>
-
 <!--- GENERIC LAMP NODE --->
 <script type="text/javascript">
     function showInput(name) {
